refactor(wp): move getPostTags to Post
